debugging 31

+++ components/Dashboard/SpecialityCard.tsx
@@ -7,6 +7,17 @@ import Link from 'next/link'
 import React from 'react'
 import toast from 'react-hot-toast'
 import { Speciality } from "@prisma/client"
+import {
+    AlertDialog,
+    AlertDialogAction,
+    AlertDialogCancel,
+    AlertDialogContent,
+    AlertDialogDescription,
+    AlertDialogFooter,
+    AlertDialogHeader,
+    AlertDialogTitle,
+    AlertDialogTrigger,
+  } from "@/components/ui/alert-dialog"
 
 export default function SpecialityCard({
     speciality
@@ -24,12 +35,29 @@ export default function SpecialityCard({
     >
     <h2>{speciality.title}</h2>
     <div className="flex">
-        <Link className='text-blue-600' href={`/dashboard/services/update/${speciality.slug}`}>
+        <Link className='text-blue-600' href={`/dashboard/speciality/update/${speciality.slug}`}>
             <Pencil className='w-4 h-4'/>
         </Link>
-        <button onClick={()=>handleDelete(speciality.id)} className='text-red-600'>
+        <AlertDialog>
+          <AlertDialogTrigger asChild>
+          <button onClick={()=>handleDelete(speciality.id)} className='text-red-600'>
             <Trash className='w- h-4'/>
         </button>
+          </AlertDialogTrigger>
+          <AlertDialogContent>
+            <AlertDialogHeader>
+              <AlertDialogTitle className='text-red-600'>Are you absolutely sure?</AlertDialogTitle>
+              <AlertDialogDescription>
+                This action cannot be undone. This will permanently delete your
+                speciality and remove your data from our servers.
+              </AlertDialogDescription>
+            </AlertDialogHeader>
+            <AlertDialogFooter>
+              <AlertDialogCancel>Cancel</AlertDialogCancel>
+              <AlertDialogAction onClick={()=> handleDelete(speciality.id)}>Delete</AlertDialogAction>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialog>
     </div>
   </div>
   )
